succed feature order

+++ challenge4/demo/src/main/java/com/challenge4/demo/model/OrderDetail.java
@@ -28,6 +28,6 @@ public class OrderDetail {
 	private Product product;
 
 	private Integer quantity;
-	private Double totalPrice;
+	private long totalPrice;
 
 }
